Keep Python filename consistent  across the repo (#464)
