+++ setup.cfg
@@ -136,7 +136,7 @@ include =
 [flake8]
 enable-extensions=G
 exclude = venv/*,tox/*,specs/*
-ignore = E123,E128,E266,E402,W503,E731,W601
+ignore = E123,E128,E266,RST303,E402,W503,E731,W601
 max-line-length = 119
 # The following are needed for flake8-rst-docstrings to make it compatible with Sphinx
 rst-roles =
