+++ tests/amazon/aws/hooks/test_s3_hooks.py
@@ -9,7 +9,7 @@ from aiobotocore.session import ClientCreatorContext
 from airflow.models.connection import Connection
 from botocore.exceptions import ClientError
 
-from astronomer.providers.amazon.aws.hooks.base_aws_async import AwsBaseHookAsync
+from astronomer.providers.amazon.aws.hooks.base_aws import AwsBaseHookAsync
 from astronomer.providers.amazon.aws.hooks.s3 import S3HookAsync
 
 
@@ -21,7 +21,7 @@ async def test_aws_base_hook_async_get_client_async_without_get_connection():
     assert isinstance(response, ClientCreatorContext)
 
 
-@mock.patch("astronomer.providers.amazon.aws.hooks.base_aws_async.AwsBaseHookAsync.get_connection")
+@mock.patch("astronomer.providers.amazon.aws.hooks.base_aws.AwsBaseHookAsync.get_connection")
 @pytest.mark.asyncio
 async def test_aws_base_hook_async_get_client_async_with_get_connection(mock_connection):
     aws_base_hook_async_obj = AwsBaseHookAsync(client_type="S3", resource_type="S3")
@@ -30,7 +30,7 @@ async def test_aws_base_hook_async_get_client_async_with_get_connection(mock_con
     assert isinstance(response, ClientCreatorContext)
 
 
-@mock.patch("astronomer.providers.amazon.aws.hooks.base_aws_async.AwsBaseHookAsync.get_connection")
+@mock.patch("astronomer.providers.amazon.aws.hooks.base_aws.AwsBaseHookAsync.get_connection")
 @pytest.mark.asyncio
 async def test_aws_base_hook_async_get_client_async_with_aws_secrets(mock_get_connection):
     mock_conn = Connection(extra=json.dumps({"aws_access_key_id": "", "aws_secret_access_key": ""}))
