+++ tests/apache/livy/operators/test_livy.py
